refactor: move to HttpCache
